+++ packages/interactions/route-prefetch/types/index.d.ts
@@ -1,3 +1,7 @@
 import { Interaction } from "@curi/core";
+export interface PrefetchType {
+    initial?: boolean;
+    every?: boolean;
+}
 declare function prefetchRoute(): Interaction;
 export default prefetchRoute;